refactor: improve readability of ModuleAvailabilityChecker
